[Fix] BudgetDetail 과정 List 최신화시 돈이 Int 범위 초과시 오류 수정

+++ app/src/main/java/kr/sparta/tripmate/ui/viewmodel/budget/detail/statistics/BudgetStatisticsViewModel.kt
@@ -35,8 +35,8 @@ class BudgetStatisticsViewModel(budgetNum: Int, repository: BudgetRepository) : 
                         }
                     }
 
-                    var currentMoney = budget.money
-                    fun reduceCurrentMoney(reduce: Int): Int {
+                    var currentMoney = budget.money.toLong()
+                    fun reduceCurrentMoney(reduce: Int): Long {
                         currentMoney -= reduce
                         return currentMoney
                     }
